==> 03_textos/TALLER/06 - GeneraDataset.ipynb <==
from requests import get
from bs4 import BeautifulSoup
import pandas as pd
import string                              
from nltk.corpus import stopwords          
from nltk.stem import PorterStemmer        
from nltk.tokenize import TweetTokenizer 

stopwords_english = stopwords.words('english')

tokenizer = TweetTokenizer(preserve_case = False, 
                           strip_handles = True,
                           reduce_len    = True)
stemmer = PorterStemmer()

webs    = pd.ExcelFile("Webs.xls")
df      = webs.parse("Categorias")
textos = pd.DataFrame(columns=['Texto','Longitud' ,'Clase'])
for i in df.index:
    url       = df["URL"][i]
    pagina    = get(url)
    html      = BeautifulSoup(pagina.text,"html.parser")
    parrafos  = [x.text.strip() for x in html.findAll("p") if x.text != "\n"]
    for p in parrafos:
        if 15 < len(p.split()) < 150:
            texto_tokens = tokenizer.tokenize(p)
            texto_clean  = [ word for word in texto_tokens if
                            word not in stopwords_english and
                            word not in string.punctuation]
            texto_stem   = [stemmer.stem(word) for word in texto_clean]
            textos = textos.append({'Texto': texto_stem, 'Longitud': len(texto_stem),'Clase':df["CÓDIGO"][i]}, ignore_index=True)

textosBack = textos

textos

textos.tail(500)

textos.describe()

textos.groupby(by=["Clase"]).count()

textos =  textos.sort_values('Longitud',ascending = True)
textos.tail(500)

textos = textos.drop(textos[textos['Longitud'] < 20].index)
textos.groupby(by=["Clase"]).count()

# textos.to_csv("Historia.csv")

textos = textos.sort_index()

textos.head(100)

textos20 = pd.DataFrame(columns=['Texto','Longitud' ,'Clase'])
for i in textos.index:    
    largo = textos.loc[i,'Longitud']    
    # print(i)    
    # print(largo)
    partes = largo // 20
    indice = 0
    texto = textos.loc[i,'Texto']
    for k in range(partes):
        texto20 = []
        for j in range(20):
             texto20.append(texto[20 * k + j])
        textos20 = textos20.append({'Texto': texto20, 'Longitud': len(texto20),'Clase':textos["Clase"][i]}, ignore_index=True)        

textos

textos20

# textos20.to_csv("Historia20.csv")

textos20.groupby(by=["Clase"]).count()

textos20.describe()

textosLargo = textos.drop("Texto", axis = 1)
textosLargo

textosLargo.describe()

textos.groupby(by=["Longitud"]).count().head(50)

textos["partesde20"] = textos["Longitud"] // 20

textos.sample(10)

textos["partesde20"].sum()

textos.groupby(by=["partesde20"]).count()
