==> 03_textos/TALLER/10 - ModeloTextos.ipynb <==
from google.colab import drive
drive.mount('/content/drive/') # montar google drive como filesystem

import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split

data=pd.read_csv("/content/drive/MyDrive/DEEP LEARNING - ESPECIALIZACION  -COURSERA/NLP/Historia.csv")# cargar un csv a dataframe de pandas

data.head()

textos = data.iloc[:,1:]
textos

X = textos.iloc[:,2:].values

type(X)

X

Y = textos.iloc[:,1].values 

type(Y)

Y

xtr,xtst,ytr,ytst=train_test_split(X,Y,test_size=.2,random_state=564)

xtrn = xtr / 1000

xtstn = xtst / 1000

len(xtr)

len(ytr)

len(xtst)

len(ytst)

xtstn.shape[1:]

from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Dense

modelo = Sequential()

modelo.add(Dense(128,input_shape = xtstn.shape[1:]))

modelo.add(Dense(6,activation="softmax"))

modelo.compile(optimizer = "sgd", loss = "sparse_categorical_crossentropy", metrics =["acc"] )

modelo.fit(xtrn,ytr,epochs=240,batch_size=105,validation_split=.1)

modelo.evaluate(xtstn,ytst)

xtst

xtst[0],ytst[0]

modelo.predict(xtst[0].reshape(1, -1)),ytst[0]

np.argmax(modelo.predict(xtst[0].reshape(1, -1))) == ytst[0]

len(xtst)

aciertos = 0
for i in range(len(xtst)):
     if np.argmax(modelo.predict(xtst[i].reshape(1, -1))) == ytst[i]:
       aciertos +=1
print("aciertos: ",aciertos)       

print("Accuracy: ", aciertos / len(xtst))

modelo.save_weights("/content/drive/MyDrive/DEEP LEARNING - ESPECIALIZACION  -COURSERA/NLP/Historia.h5")

modelo.save("/content/drive/MyDrive/DEEP LEARNING - ESPECIALIZACION  -COURSERA/NLP/MHistoriaFreq.h5")
