==> 03_textos/TALLER/11 - ConsumirModeloEntrenado .ipynb <==
import pandas as pd
import numpy as np
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Dense
import keras as K
import pickle

from google.colab import drive
drive.mount('/content/drive/') # montar google drive com

textos = pd.read_csv("/content/drive/MyDrive/DEEP LEARNING - ESPECIALIZACION  -COURSERA/NLP/Historia.csv")

textos.head()

textos = textos.iloc[:,1:]
textos

X = textos.iloc[:,2:].values

Y = textos.iloc[:,1].values

Xn = X / 1000

len(Xn)

modelo = Sequential()
modelo.add(Dense(128,input_shape = Xn.shape[1:]))
modelo.add(Dense(6,activation="softmax"))
modelo.compile(optimizer = "sgd", loss = "sparse_categorical_crossentropy", metrics =["acc"] )


modelo.load_weights('/content/drive/MyDrive/DEEP LEARNING - ESPECIALIZACION  -COURSERA/NLP/Historia.h5')

np.argmax(modelo.predict(Xn[10122].reshape(1, -1))),Y[10122]

# url = "https://www.history.com/topics/cold-war/cuban-missile-crisis"
# url = "https://courses.lumenlearning.com/suny-hccc-worldhistory/chapter/the-seven-kings/"
# url = "https://heritagecalling.com/2019/02/22/a-brief-introduction-to-prehistoric-caves/"
# url = "https://www.ancient.eu/Medieval_Castle/" FALLO
# url = "https://www.britannica.com/event/Battle-of-Tours-732"
# url = "https://www.historytoday.com/archive/1683-siege-vienna"
# url = "https://www.sciencedirect.com/topics/social-sciences/early-modern-times"
# url = "https://www.sparknotes.com/lit/3musk/section15/"  LA PEGO Y NO SE PORQUE
# url = "https://marcopologencana.wordpress.com/2019/06/17/the-crazy-years-between-the-wars/"
# url = "https://www.nationalgeographic.com/news/2006/9/lucy-facts-on-early-human-ancestor/"
# url = "https://www.researchgate.net/publication/236765312_Spartacus_Film_and_History_and_Troy_From_Homer's_'Iliad'_to_Hollywood_Epic_review"
# url = "https://www.thespanishblog.com/2011/06/history-of-spain-the-carthaginians/"
url = "https://www.history.com/news/prehistoric-ages-timeline"




import pickle
def save_obj(obj, name ):
       with open(name + '.pkl', 'wb') as f:
           pickle.dump(obj, f, pickle.HIGHEST_PROTOCOL)

def load_obj(name ):
       with open('/content/drive/MyDrive/DEEP LEARNING - ESPECIALIZACION  -COURSERA/NLP//' + name + '.pkl', 'rb') as f:

           return pickle.load(f)

Vocabulario = load_obj("Vocabulario")
frecuencias = load_obj("Frecuencias")
clases_df = load_obj("clases_df")

clases_df

from requests import get # la vamos a usar para solicitar la página mediante su URL 
from bs4 import BeautifulSoup # biblioteca para web scraping
import pandas as pdf
import string                              
from nltk.corpus import stopwords # para remover palabras que no agregan significado          
from nltk.stem import PorterStemmer # para llevar las palabras a su forma raíz        
from nltk.tokenize import TweetTokenizer # para transformar cada palabra en un token
from IPython.display import clear_output
from collections import defaultdict
import numpy as np
import nltk
nltk.download('stopwords')

def etiquetas(palabras,frecuencias,clases):
    import numpy as np
    x = np.zeros((1, 6))
    for palabra in palabras:
        for t in clases:
            x[0,t] += frecuencias.get((palabra, t),0)
    x[0,0]=1
    return x

stop        = stopwords.words('english')
tokenizer   = TweetTokenizer(preserve_case = False, strip_handles = True, reduce_len = True)
stemmer     = PorterStemmer()
textosTest  = pd.DataFrame(columns = ['Texto'])
pagina    = get(url)
html      = BeautifulSoup(pagina.text,"html.parser")
parrafos  = [ x.text.strip() for x in html.findAll("p") if x.text != "\n" ]
for p in parrafos:
    palabras = p.split()
    if 15 <= len(palabras):
       clear_output( wait = True )
       print("procesando {}".format(url))
       texto_tokens = tokenizer.tokenize(p)
       texto_clean  = [ word for word in texto_tokens if word not in stop and word not in string.punctuation ]
       texto_stem   = [ stemmer.stem(word) for word in texto_clean ]
       textosTest   = textosTest.append({'Texto' : texto_stem}, ignore_index = True)

vocabularioTest = set([x for m in textosTest.index for x in textosTest.loc[m].Texto ])                



for i in clases_df:
    textosTest[f"Clase{i}"] = textosTest.Texto.apply(lambda x : etiquetas(x,frecuencias,clases_df)[0,i])
print("dataset terminado")
print(textosTest.describe())

textosTest

XTest = textosTest.iloc[:,1:].values

XTest

XTestn = XTest / 1000

XTestn


modelo.predict(XTestn[0].reshape(1, -1))

np.argmax(modelo.predict(XTestn[0].reshape(1, -1)))

len(vocabularioTest)

resultados = {}
for xT in range (len(XTestn)):
  print(np.argmax(modelo.predict(XTestn[xT].reshape(1, -1))),end = " ")
  r = np.argmax(modelo.predict(XTestn[xT].reshape(1, -1)))
  resultados[r] = resultados.get(r, 0) + 1
print()  
resultados

print(len(frecuencias))
for i in  frecuencias.keys():
    print(i)
    print(frecuencias[i])

textos = textos.drop("Longitud", axis = 1)
textos

X = textos.iloc[:,2:].values
