==> 03_textos/TALLER/12 - Crea_dataset_confusas.ipynb <==
from requests import get # la vamos a usar para solicitar la página mediante su URL 
from bs4 import BeautifulSoup # biblioteca para web scraping
import pandas as pd
import string                              
from nltk.corpus import stopwords # para remover palabras que no agregan significado          
from nltk.stem import PorterStemmer # para llevar las palabras a su forma raíz        
from nltk.tokenize import TweetTokenizer # para transformar cada palabra en un token
from IPython.display import clear_output
from collections import defaultdict
import numpy as np

import pickle
def save_obj(obj, name):
       with open(name + '.pkl', 'wb') as f:
           pickle.dump(obj, f, pickle.HIGHEST_PROTOCOL)

def load_obj(name):
       with open(name + '.pkl', 'rb') as f:
           return pickle.load(f)

def etiquetas(palabras,frecuencias,clases):
    import numpy as np
    x = np.zeros((1, 6))
    for palabra in palabras:
        for t in clases:
            x[0,t] += frecuencias.get((palabra, t),0)
    x[0,0]=1
    return x

stop        = stopwords.words('english')
tokenizer   = TweetTokenizer(preserve_case = False, strip_handles = True, reduce_len = True)
stemmer     = PorterStemmer()
webs        = pd.ExcelFile("WEBs.xls")
df          = webs.parse("Categorias")
textos      = pd.DataFrame(columns = [ 'Texto' , 'Clase'])
clases_df   = set(df["CÓDIGO"].value_counts().index)
frecuencias = defaultdict(int)
for i in df.index:
    url       = df["URL"][i]
    pagina    = get(url)
    html      = BeautifulSoup(pagina.text,"html.parser")
    parrafos  = [ x.text.strip() for x in html.findAll("p") if x.text != "\n" ]
    for p in parrafos:
        palabras = p.split()
        if 15 <= len(palabras):
            clear_output( wait = True )
            print("procesando {}".format(url))
            print(f"{i} de {len(df)}")
            texto_tokens = tokenizer.tokenize(p)
            texto_clean  = [ word for word in texto_tokens if word not in stop and word not in string.punctuation ]
            texto_stem   = [ stemmer.stem(word) for word in texto_clean ]
            for t in range(len(texto_stem)//20):
                j = 20 * t
                palabras  = texto_stem[ j : 20 * (t+1) ] 
                textos   = textos.append({'Texto' : palabras, 'Clase' : df["CÓDIGO"][i] }, ignore_index = True)

vocabulario = set([x for m in textos.index for x in textos.loc[m].Texto ])                
clases_df   = set(df["CÓDIGO"].value_counts().index)

for i in textos.index:
    for palabra in textos.loc[i].Texto:
        frecuencias[(palabra, textos.loc[i].Clase)] += 1
        clear_output( wait = True )
        print(f"Creando Frecuencias {i} de {len(textos)}")

for i in clases_df:
    textos[f"Clase{i}"] = textos.Texto.apply(lambda x : etiquetas(x,frecuencias,clases_df)[0,i])
print("dataset terminado")
print(textos.describe())

textos

textos.describe()

textos.info()

textos.to_csv("Historia.csv")

import pandas as pd
textos = pd.read_csv("Historia.csv")
textos = textos.iloc[:,1:]
textos

textos.groupby(by=["Clase"]).count()

textos.sample(10)

len(vocabulario)

len(frecuencias)

clases_df

save_obj(vocabulario, "Vocabulario")

save_obj(frecuencias, "Frecuencias")

save_obj(clases_df, "clases_df")

textos.sample(10)

int(np.argmax(textos.iloc[3337,-5:].values)) + 1

int(textos.iloc[14317,1:2].values)

int(np.argmax(textos.iloc[1621,-5:].values)) + 1 == int(textos.iloc[1621,1:2].values)

textos.sample(10)

i = 3
int(np.argmax(textos.iloc[i,-6:].values)) == int(textos.iloc[i,1:2].values)

int(np.argmax(textos.iloc[i,-5:].values) + 1)

textos.iloc[i,-6:].values

textos["Clase1"][0]

# textos = textos.iloc[:10,:]
textos

textos["FreqOK"] = None
textos

for i in textos.index:
    textos["FreqOK"][i] = int(np.argmax(textos.iloc[i,-6:-1].values)) + 1 == int(textos.iloc[i,1:2].values)
    # print((np.argmax(textos.iloc[i,-6:-1].values)) + 1 == int(textos.iloc[i,1:2].values))
textos    

textos.sample(10)

textos.sort_values(by=['FreqOK'])
textos

textos.groupby(by=["FreqOK"]).count()

textosF = textos[textos['FreqOK'] == False]
textosF

textosF.groupby(by=["Clase"]).count()

textosF1 = textosF[textosF['Clase'] == 1]
textosF1

claseMax = np.argmax(textosF1.iloc[0,-6:-1].values) + 1
claseMax

textosF1["ClasMax"] = None
textosF1

indice = 0
for i in textosF1.index:
    textosF1["ClasMax"][i] = np.argmax(textosF1.iloc[indice,-7:-2].values) + 1
    indice +=1
textosF1.head(11)    

textosF1.sample(10)    

frecuencias = load_obj("Frecuencias")

vocabulario = load_obj("Vocabulario")

claseTrue = 1
clases = 5
palabrasConfusas = []
# indice = 0
totalPalabras = 0
for i in textosF1.index:
    """
    if indice > 1:
        break"""
    
    w = 0        
    for word in textosF1.loc[i,"Texto"].split():
        freqWord = []
        for k in range(clases + 1):
            freqWord.append(0)
        if w == 0:
            word = word[2:-2]
        else:   
            word = word[1:-2]
        w += 1                  
        if word in vocabulario:
            for k in range(1, len(freqWord)):
                print(word,"clase ",k,frecuencias[(word,k)])
                freqWord[k] = frecuencias[(word,k)]
        print(freqWord)    
        print(np.argmax(freqWord))
        print(np.argmax(freqWord) == claseTrue)        
        if np.argmax(freqWord) != claseTrue:
            palabrasConfusas.append(word)
        totalPalabras += w        
    indice += 1
print("palabras",totalPalabras)
print("confusas",len(palabrasConfusas))
palabrasConfusas

claseTrue = 1
clases = 5

for word in palabrasConfusas:
    freqWord = []
    for k in range(clases + 1):
        freqWord.append(0)
    if word in vocabulario:
        print()                                  
        for i in range(1,6):
            print(word,"clase ",i,frecuencias[(word,i)]) 
            freqWord[i] = frecuencias[(word,i)]
        print(freqWord)    
        print(np.argmax(freqWord))
        print(np.argmax(freqWord) == claseTrue)
palabrasConfusas        

if 'bronz' in vocabulario:
        print("está")

print(frecuencias)

len(palabrasConfusas)

for word in palabrasConfusas:
    if word in vocabulario:
        print()                                  
        for i in range(1,6):
            print(word,"clase ",i,frecuencias[(word,i)]) 
            freqWord[i] = frecuencias[(word,i)]
        print(freqWord)    
        print(np.argmax(freqWord))
        print(np.argmax(freqWord) == claseTrue)
        print("Maximo", freqWord[np.argmax(freqWord)],"Clase",np.argmax(freqWord))
palabrasConfusas        

confusas = pd.DataFrame(columns = ['Palabra' , 'Clase','FrecClase','ClaseMax','FrecMax'])

for word in palabrasConfusas:
    if word in vocabulario: 
        freqWord = []
        for k in range(clases + 1):
            freqWord.append(0)
        for i in range(1,6):
            freqWord[i] = frecuencias[(word,i)]
        confusas = confusas.append({'Palabra' : word, 'Clase' : claseTrue ,'FrecClase' : freqWord[claseTrue], 'ClaseMax' : np.argmax(freqWord),'FrecMax' : freqWord[np.argmax(freqWord)] }, ignore_index = True)        
confusas

confusas.to_csv("Palabras_Confusas.csv")

import pandas as pd
confusas = pd.read_csv("Palabras_Confusas.csv")
confusas

confusas = confusas.iloc[:,1:]
confusas

confusas["Porcentaje"] = 100 * confusas["FrecClase"] / confusas["FrecMax"]
confusas

confusas["Decenas"] = confusas["Porcentaje"] // 10
confusas

confusas.groupby(by=["Decenas"]).count()

confusas_clase1 = []
for i in confusas.index:
    if confusas["Decenas"][i] < 2:
        confusas_clase1.append(confusas["Palabra"][i])
len(confusas_clase1)        

confusas_clase1

frecuencias

claseTrue = 1
clases = 5
cont = 0
for word in confusas_clase1:
    print()
    cont +=1
    print(word)
    for clase in range(1,clases + 1):
        print(frecuencias.get((word,clase),0))
        

claseTrue = 1
clases = 5
frecuenciasMod = frecuencias.copy()
for word in confusas_clase1:
    frecuenciasMod[word,claseTrue] = 1

claseTrue = 1
clases = 5
for word in confusas_clase1:
    print()
    print()
    print(word)
    for clase in range(1,clases + 1):
        print(frecuencias.get((word,clase),0), end = " ")
    print()    
    for clase in range(1,clases + 1):
        print(frecuenciasMod.get((word,clase),0), end = " ")
