==> 03_textos/TALLER/16 - ObtenerPifiadas .ipynb <==
import pandas as pd
import numpy as np
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Dense
import keras as K
import pickle

from google.colab import drive
drive.mount('/content/drive/') # montar google drive com

textos = pd.read_csv("/content/drive/MyDrive/NLP Specialization/MODELO/H03 - HistoriaFreq.csv")

textos.head()

textos = textos.drop(labels = "Texto", axis = 1 )
textos

XTEXTOS = textos.drop(labels = "Clase", axis = 1 ).values


X = textos.iloc[:,2:].values


Y = textos.iloc[:,1].values

X

Xn = X / 1000

len(Xn)

modelo = Sequential()
modelo.add(Dense(128,input_shape = Xn.shape[1:]))
modelo.add(Dense(6,activation="softmax"))
modelo.compile(optimizer = "sgd", loss = "sparse_categorical_crossentropy", metrics =["acc"] )


modelo.load_weights('/content/drive/MyDrive/NLP Specialization/MODELO/Historia.h5')

np.argmax(modelo.predict(Xn[5125].reshape(1, -1))),Y[5125]

clases_df

modelo.predict(Xn[0].reshape(1, -1))

np.argmax(modelo.predict(Xn[0].reshape(1, -1)))

resultados = {}
for xT in range (len(Xn)):
  print(np.argmax(modelo.predict(Xn[xT].reshape(1, -1))),end = " ")
  r = np.argmax(modelo.predict(Xn[xT].reshape(1, -1)))
  resultados[r] = resultados.get(r, 0) + 1
print()  
resultados

textos.head()

textos["Predicciones"] = 0

textos

cont = 0
for xT in range (len(Xn)):
  cont += 1
  if cont > 20:
    break
  textos["Predicciones"][xT] = np.argmax(modelo.predict(Xn[xT].reshape(1, -1)))
textos

for xT in range (len(Xn)):
  textos["Predicciones"][xT] = np.argmax(modelo.predict(Xn[xT].reshape(1, -1)))
textos

textos.to_csv("/content/drive/MyDrive/NLP Specialization/MODELO/H04 - Predicciones.csv")
