==> 03_textos/TALLER/22_Clase_Textos_ENZO.ipynb <==
from requests import get # la vamos a usar para solicitar la página mediante su URL 
from bs4 import BeautifulSoup # biblioteca para web scraping
import pandas as pd
import string                              
from nltk.corpus import stopwords # para remover palabras que no agregan significado          
from nltk.stem import PorterStemmer # para llevar las palabras a su forma raíz        
from nltk.tokenize import TweetTokenizer # para transformar cada palabra en un token
from IPython.display import clear_output
from collections import defaultdict
import numpy as np
import nltk
nltk.download('stopwords')

class Textos:
    
    def __init__(self, clases, webs):
        self.clases = clases
        self.lista_dataset = []
        self.crear_dataframe(clases, webs)

    def crear_dataframe(self, clases, webs):
        self.lista_dataset.append('TextoCrudo')
        self.lista_dataset.append('TextoLimpio')

        for i in range(1, clases + 1):
            self.lista_dataset.append('Clase' + str(i))

        self.lista_dataset.append('Predicción')
        self.lista_dataset.append('CLASE')
        self.dataframe = pd.DataFrame(columns=self.lista_dataset)
        self.webs = pd.ExcelFile(webs)
        df = self.webs.parse(self.webs.sheet_names[0])
        self.__clases_df = set(df["CÓDIGO"].value_counts().index)

        for i in df.index:
            print(f"Creando Dataframe {i+1} de {len(df.index)}")
            url = df["URL"][i]
            req = get(url)
            statusCode = req.status_code
            html = BeautifulSoup(req.text, "html.parser")

            if statusCode == 200:
                paragraphs = html.select("p")
                cont = 0

                for para in paragraphs:
                    if len(para.text) < 15:
                        continue
                    cont += 1
                    self.dataframe = self.dataframe.append({'TextoCrudo': para.text, 'CLASE': df["CÓDIGO"][i]}, ignore_index=True)

                clear_output(wait=True)

    def limpiar(textoCrudo):
        for i in range(100):
            textoCrudo = textoCrudo.replace("[" + str(i) + "]", "")
        
        for i in ["a","b","c"]:
            textoCrudo = textoCrudo.replace("[" + i + "]", "")
        
        stopwords_english = stopwords.words('english')
        stop        = stopwords.words('english')
        tokenizer = TweetTokenizer(preserve_case=False, strip_handles=True, reduce_len=True)
        stemmer = PorterStemmer()    
        texto_tokens = tokenizer.tokenize(textoCrudo)
        sufijos_fecha = ["bc","ac"]

        for i in range(len(texto_tokens.copy())):
            if i > len(texto_tokens) - 1:
                break
            if i !=0:
                if texto_tokens[i] in sufijos_fecha and texto_tokens[i - 1].isnumeric():
                    texto_tokens[i] = texto_tokens[i - 1] + texto_tokens[i]     
                    texto_tokens.pop(i - 1)
                           
        texto_clean  = [ word for word in texto_tokens if word not in stop and word not in string.punctuation]  
        texto_stem   = [stemmer.stem(word) for word in texto_clean] 
        return texto_stem       
      

clases = 5
webs = "webs.xlsx"
textos = Textos(clases, webs)

textos.dataframe

textos.dataframe.to_csv("Historia.csv")

dataframe = pd.read_csv("Historia.csv")
dataframe = dataframe.iloc[:,1:]
dataframe

dataframe.iloc[4029,0]

dataframe.describe()

crudo = "10 bc bcc Hola 10 bcc chau 5 bc 555 bc bcc 10 7000 ac"
limpio = Textos.limpiar(crudo)
print(type(limpio))
limpio

dataframe["LTLimpio"], dataframe["LTCrudo"] = 0, 0

for i in dataframe.index:
    dataframe["TextoLimpio"][i] = Textos.limpiar(dataframe["TextoCrudo"][i])
    dataframe["LTCrudo"][i] = len(dataframe["TextoCrudo"][i])
    dataframe["LTLimpio"][i] = len(dataframe["TextoLimpio"][i])

dataframe  

dataframe["TextoLimpio"][6063]

conta = 0
for i in dataframe.index:
    if len(dataframe["TextoLimpio"][i]) < 10:
        conta += 1 
        print(dataframe["TextoLimpio"][i]) 
print(conta)        

dataframe = dataframe[dataframe["LTLimpio"] > 7]
dataframe

dataframe.to_csv("Historia.csv")
